Simulations with Exact Means and Covariances

This function generate scenarios from elliptical distributions in which the sample moments match the populational moments.

MvnRnd(M, S, J)

Arguments

M
A 1xN matrix with the location parameter of each invariant (asset).

S
A NxN matrix with the dispersion matrix of the invariants (assets).

J
A numeric scalar with the desired number of scenarios.

Value

A J x N matrix with the simulated time-series.

References

Meucci, Attilio, Simulations with Exact Means and Covariances (June 7, 2009). Available at SSRN: https://ssrn.com/abstract=1415699 or http://dx.doi.org/10.2139/ssrn.1415699.

Attilio Meucci (2020). Simulations with Exact Means and Covariances, MATLAB Central File Exchange. Retrieved October 11, 2020.

Examples

rets  <- diff(log(EuStockMarkets))
mu    <- colMeans(rets)
sigma <- stats::cov(rets)
MvnRnd(M = mu, S = sigma, J = 50)

#>                [,1]          [,2]         [,3]          [,4]
#>  [1,] -0.0127130384 -1.045572e-02 -0.012862113 -1.898463e-03
#>  [2,]  0.0070448857 -6.251087e-03  0.006293175  9.246718e-04
#>  [3,] -0.0227159584 -1.502732e-02 -0.019613545 -1.205578e-02
#>  [4,]  0.0017884267 -5.159242e-03  0.003645726  5.083051e-04
#>  [5,]  0.0069780532 -4.219310e-03  0.012043797  3.175783e-03
#>  [6,]  0.0184316593  6.486254e-03  0.003431048  1.007073e-02
#>  [7,] -0.0048934896 -1.777308e-02 -0.013305346 -1.711326e-02
#>  [8,] -0.0057339098  8.491420e-06 -0.007704627  1.186204e-02
#>  [9,]  0.0009103128  6.718441e-03 -0.008380574 -1.593450e-03
#> [10,] -0.0051996114  2.229460e-03 -0.002215979 -4.016076e-04
#> [11,] -0.0024635471 -7.452385e-03 -0.006490622  3.524423e-03
#> [12,] -0.0023100174  3.546240e-03  0.017372738 -2.089807e-03
#> [13,]  0.0169346354  1.434759e-02  0.017390185  1.492735e-02
#> [14,] -0.0122176612 -1.398483e-02 -0.012043032 -8.018818e-03
#> [15,]  0.0064443121  1.146655e-02  0.002560629 -2.545298e-03
#> [16,] -0.0127258994 -1.373191e-02 -0.014424122 -1.230233e-02
#> [17,] -0.0020686925 -1.023488e-02 -0.004049935  3.512710e-03
#> [18,]  0.0045841070 -4.242223e-03 -0.001044426  1.656381e-03
#> [19,]  0.0045144715  3.422975e-03  0.005322910  5.087520e-03
#> [20,] -0.0015649601 -6.071815e-03 -0.006889454 -1.093726e-02
#> [21,]  0.0152170094  1.269047e-02 -0.006429843 -2.038063e-03
#> [22,]  0.0017346400  5.291736e-03  0.003198462  3.108188e-03
#> [23,] -0.0115384797 -6.945633e-03 -0.016057732  1.128767e-03
#> [24,]  0.0130246974  9.577617e-04  0.004485607  6.099631e-03
#> [25,]  0.0149491028  3.686875e-03  0.022537956  1.462786e-02
#> [26,]  0.0140171219  1.209152e-02  0.013736221  2.762433e-03
#> [27,] -0.0057408022  7.886886e-03 -0.005419067 -6.070167e-05
#> [28,]  0.0240200419  1.666312e-02  0.020487653  1.291975e-02
#> [29,] -0.0004843432  6.795041e-03 -0.002771618  3.556650e-04
#> [30,] -0.0056739697  5.855110e-03 -0.011169689 -2.311813e-03
#> [31,] -0.0171275758 -4.850454e-03 -0.002556940 -9.206762e-03
#> [32,]  0.0061975731  1.940888e-02  0.014179454  1.797723e-02
#> [33,]  0.0070379933  1.627308e-03  0.008578734 -1.099807e-02
#> [34,]  0.0003937707 -5.082642e-03  0.009254682  2.457420e-03
#> [35,]  0.0065036949 -5.936610e-04  0.003090087  1.265578e-03
#> [36,]  0.0037676306  9.088184e-03  0.007364730 -2.660452e-03
#> [37,]  0.0036141009 -1.910440e-03 -0.016498630  2.953777e-03
#> [38,] -0.0156305519 -1.271179e-02 -0.016516077 -1.406338e-02
#> [39,]  0.0135217447  1.562063e-02  0.012917140  8.882789e-03
#> [40,] -0.0051402286 -9.830755e-03 -0.001686521  3.409268e-03
#> [41,]  0.0140299829  1.536771e-02  0.015298230  1.316630e-02
#> [42,]  0.0033727760  1.187068e-02  0.004924043 -2.648740e-03
#> [43,] -0.0032800235  5.878022e-03  0.001918534 -7.924105e-04
#> [44,] -0.0032103880 -1.787176e-03 -0.004448802 -4.223550e-03
#> [45,]  0.0028690436  7.707614e-03  0.007763562  1.180123e-02
#> [46,] -0.0139129259 -1.105467e-02  0.007303951  2.902033e-03
#> [47,] -0.0004305565 -3.655936e-03 -0.002324354 -2.244217e-03
#> [48,]  0.0128425632  8.581432e-03  0.016931840 -2.647967e-04
#> [49,] -0.0117206139  6.780377e-04 -0.003611499 -5.235661e-03
#> [50,] -0.0136450193 -2.051076e-03 -0.021663848 -1.376389e-02