Simulations with Exact Means and Covariances

This function generate scenarios from elliptical distributions in which the sample moments match the populational moments.

MvnRnd(M, S, J)

Arguments

M
A 1xN matrix with the location parameter of each invariant (asset).

S
A NxN matrix with the dispersion matrix of the invariants (assets).

J
A numeric scalar with the desired number of scenarios.

Value

A J x N matrix with the simulated time-series.

References

Meucci, Attilio, Simulations with Exact Means and Covariances (June 7, 2009). Available at SSRN: https://ssrn.com/abstract=1415699 or http://dx.doi.org/10.2139/ssrn.1415699.

Attilio Meucci (2020). Simulations with Exact Means and Covariances, MATLAB Central File Exchange. Retrieved October 11, 2020.

Examples

rets  <- diff(log(EuStockMarkets))
mu    <- colMeans(rets)
sigma <- stats::cov(rets)
MvnRnd(M = mu, S = sigma, J = 50)

#>                [,1]          [,2]          [,3]          [,4]
#>  [1,] -0.0082137937 -0.0103948384 -0.0105914975 -0.0122983399
#>  [2,]  0.0045853444 -0.0062843668  0.0050519284  0.0066098322
#>  [3,] -0.0212499698 -0.0150074828 -0.0188737107 -0.0154443760
#>  [4,]  0.0010310789 -0.0051694893  0.0032635183  0.0022588932
#>  [5,]  0.0052280173 -0.0042429900  0.0111606137  0.0072209420
#>  [6,]  0.0193145053  0.0064981994  0.0038765902  0.0080300583
#>  [7,] -0.0133275274 -0.0178871985 -0.0175617178  0.0023817802
#>  [8,]  0.0054595569  0.0001599502 -0.0020556642 -0.0140113462
#>  [9,]  0.0007375608  0.0067161037 -0.0084677560 -0.0011941382
#> [10,] -0.0036097953  0.0022509721 -0.0014136525 -0.0040764228
#> [11,]  0.0017414291 -0.0073954874 -0.0043685138 -0.0061952615
#> [12,] -0.0057811200  0.0034992721  0.0156209902  0.0059335489
#> [13,]  0.0195065021  0.0143823867  0.0186881185  0.0089825532
#> [14,] -0.0122678358 -0.0139855048 -0.0120683535 -0.0079028413
#> [15,]  0.0022338127  0.0114095824  0.0004357334  0.0071871527
#> [16,] -0.0152601707 -0.0137661987 -0.0157030825 -0.0064444367
#> [17,]  0.0016464184 -0.0101846105 -0.0021750444 -0.0050746641
#> [18,]  0.0044974752 -0.0042433952 -0.0010881459  0.0018566276
#> [19,]  0.0058475677  0.0034410130  0.0059956787  0.0020061057
#> [20,] -0.0077100735 -0.0061549644 -0.0099906840  0.0032669931
#> [21,]  0.0100729791  0.0126208645 -0.0090258603  0.0098522184
#> [22,]  0.0028408449  0.0053067038  0.0037567258  0.0005512261
#> [23,] -0.0047996848 -0.0068544501 -0.0126568912 -0.0144477677
#> [24,]  0.0126254380  0.0009523593  0.0042841142  0.0070225079
#> [25,]  0.0171692190  0.0037169158  0.0236583736  0.0094961248
#> [26,]  0.0095178772  0.0120306377  0.0114656054  0.0131623100
#> [27,] -0.0032812609  0.0079201661 -0.0041778204 -0.0057458621
#> [28,]  0.0225540533  0.0166432821  0.0197478187  0.0163083462
#> [29,]  0.0002730046  0.0068052886 -0.0023894103 -0.0013949230
#> [30,] -0.0039239338  0.0058787893 -0.0102865058 -0.0063569718
#> [31,] -0.0180104218 -0.0048624001 -0.0030024823 -0.0071660881
#> [32,]  0.0146316108  0.0195229978  0.0184358257 -0.0015178100
#> [33,] -0.0041554734  0.0014758491  0.0029297722  0.0148753163
#> [34,]  0.0005665227 -0.0050803044  0.0093418640  0.0020581083
#> [35,]  0.0049138788 -0.0006151728  0.0022877605  0.0049403929
#> [36,] -0.0004373456  0.0090312867  0.0052426217  0.0070592317
#> [37,]  0.0070852035 -0.0018634727 -0.0147468822 -0.0050695788
#> [38,] -0.0182024186 -0.0127465874 -0.0178140105 -0.0081185831
#> [39,]  0.0135719193  0.0156213041  0.0129424614  0.0087668114
#> [40,] -0.0009297292 -0.0097737831  0.0004383746 -0.0063231826
#> [41,]  0.0165642542  0.0154019980  0.0165771905  0.0073084068
#> [42,] -0.0003423349  0.0118204099  0.0030491524  0.0059386343
#> [43,] -0.0031933917  0.0058791945  0.0019622538 -0.0009926574
#> [44,] -0.0045434842 -0.0018052137 -0.0051215707 -0.0011421355
#> [45,]  0.0090141570  0.0077907638  0.0108647920 -0.0024030229
#> [46,] -0.0087688956 -0.0109850652  0.0098999682 -0.0089882483
#> [47,] -0.0015367615 -0.0036709045 -0.0028826178  0.0003127440
#> [48,]  0.0061037683  0.0084902494  0.0135309992  0.0153117379
#> [49,] -0.0113213545  0.0006834400 -0.0034100062 -0.0061585377
#> [50,] -0.0158651355 -0.0020811165 -0.0227842656 -0.0086321546